move prompt_info to internal
